+++ src/curvaturevelocityoperator.hh
@@ -115,7 +115,7 @@ class CurvatureVelocityOperator:public Operator<DomainFunctionImp,RangeFunctionI
 
       // extract the associated bulk intersection
       auto intersectionIt(bulkgridpart_.ibegin(bulkEntity));
-      while(intersectionIt->indexInInside()!=faceLocalIdx)
+      while(static_cast<std::size_t>(intersectionIt->indexInInside())!=faceLocalIdx)
         ++intersectionIt;
       const auto intersection(*intersectionIt);
